getActionType.js
