+++ packages/client-redux/test/client-redux.test.js
@@ -1,6 +1,5 @@
 import Promise from 'bluebird';
 import reduxStellar from '../src';
-import { getActionType } from '../src/getActionType';
 import { mockAction, mockRef, mockStellarSocket } from './mocks';
 
 describe('client-redux', () => {
@@ -66,35 +65,5 @@ describe('client-redux', () => {
     middleware(mockRef)(mockNext)(action);
     expect(mockNext)
           .toHaveBeenLastCalledWith({ type: action.type, payload: { promise: expect.any(Promise), data: action.payload } });
-  });
-
-  it('should return action type - 3 parameters', () => {
-    function testFn(id, options, dispatch = () => null) {
-      return dispatch({
-        resource: 'a',
-        method: 'b',
-        path: 'c',
-        payload: id,
-        options,
-      });
-    }
-    const x = getActionType(testFn);
-
-    expect(x).toEqual('a:b:c');
-  });
-
-  it('should return action type - 4 parameters', () => {
-    function testFn({ id }, { bla }, options, dispatch = () => null) {
-      return dispatch({
-        resource: 'a',
-        method: 'b',
-        path: 'c',
-        payload: { id, bla },
-        options,
-      });
-    }
-    const x = getActionType(testFn);
-
-    expect(x).toEqual('a:b:c');
   });
 });
